fix: decode to values up to 255, not 256 - step
This commit expands the color gamut.
It does not increase color accuracy in the abstract
(it is still one value to four or eight),
but it may increase color accuracy in practice.
Images as a whole most likely have more 255 than 248 or 252 color values.

Use tables for encoding and decoding.
It makes the mapping of values visible and easier to understand.
It does not seem to reduce performance on large files.